organizing discover backend into algorithms, api, model, worker
