+++ python/backend/backend/discover/algorithms/direct_lingam.py
@@ -6,8 +6,11 @@ import numpy as np
 from castle.algorithms import DirectLiNGAM
 from networkx.readwrite import json_graph
 
-from backend.discover.algorithms.base import CausalDiscoveryRunner, CausalGraph
-from backend.discover.base_payload import CausalDiscoveryPayload
+from backend.discover.algorithms.commons.base_runner import (
+    CausalDiscoveryRunner,
+    CausalGraph,
+)
+from backend.discover.model.causal_discovery import CausalDiscoveryPayload
 
 
 class DirectLiNGAMPayload(CausalDiscoveryPayload):
