+++ python/backend/backend/discover/algorithms/notears.py
@@ -1,8 +1,11 @@
 from causalnex.structure.notears import from_pandas
 from networkx.readwrite import json_graph
 
-from backend.discover.algorithms.base import CausalDiscoveryRunner, CausalGraph
-from backend.discover.base_payload import CausalDiscoveryPayload
+from backend.discover.algorithms.commons.base_runner import (
+    CausalDiscoveryRunner,
+    CausalGraph,
+)
+from backend.discover.model.causal_discovery import CausalDiscoveryPayload
 
 
 class NotearsPayload(CausalDiscoveryPayload):
